Added the ForgotPassword Pages and Link Between Homepages and The Rest
