+++ src/App.tsx
@@ -7,11 +7,12 @@ import {
 
 import LoginPage from "./components/LoginPage";
 import ForgottenPasswordPage from "./components/ForgottenPasswordPage";
+import PasswordQuestionPage from "./components/ForgottenPasswordPage/PasswordQuestionPage";
 
-import NavBar from "./Compoents/NavBar";
-import Footer from "./Compoents/Footer";
-import Slider from "./Compoents/Slider";
-import Dashboard from "./Compoents/dashboard";
+import NavBar from "./components/HomePage/NavBar";
+import Footer from "./components/HomePage/Footer";
+import Slider from "./components/HomePage/Slider";
+import Dashboard from "./components/Dashboard";
 
 function App() {
   return (
@@ -19,10 +20,13 @@ function App() {
       <Routes>
         <Route path="/" element={<LoginPage />} />
         <Route path="/forgotpassword" element={<ForgottenPasswordPage />} />
+        <Route path="/home" element={<Dashboard />} />
+        <Route path="/password_questions" element={<PasswordQuestionPage />} />
         <Route path="*" element={<Navigate to="/" />} />
       </Routes>
     </Router>
     //guukug
+  );
 }
 
 export default App;
