--- src/Compoents/Dashboard.tsx
@@ -1,18 +1,0 @@
-// Dashboard.js
-
-import React from "react";
-import NavBar from "./NavBar";
-import Footer from "./Footer";
-import Slider from "./Slider";
-
-function Dashboard() {
-  return (
-    <div>
-      <NavBar />
-      <Footer />
-      <Slider />
-    </div>
-  );
-}
-
-export default Dashboard;
