--- src/Compoents/Slider.css
@@ -1,10 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  background-image: url("../Compoents/back.png"); /* Your base64-encoded image here */
-  background-size: 200vh;
-  background-position: center;
-  background-repeat: no-repeat;
-
-  height: 100vh;
-}
