+++ src/components/Dashboard.tsx
@@ -1,0 +1,18 @@
+// Dashboard.js
+
+import React from "react";
+import NavBar from "./HomePage/NavBar";
+import Footer from "./HomePage/Footer";
+import Slider from "./HomePage/Slider";
+
+function Dashboard() {
+  return (
+    <div>
+      <NavBar />
+      <Footer />
+      <Slider />
+    </div>
+  );
+}
+
+export default Dashboard;
