+++ src/components/ForgottenPasswordPage/PasswordQuestionPage.tsx
@@ -1,0 +1,12 @@
+import BackgroundFPassword from "./BackgroundFPassword";
+import InputAnswer from "./InputAnswer";
+
+const PasswordQuestionPage = () => {
+  return (
+    <BackgroundFPassword>
+      <InputAnswer></InputAnswer>
+    </BackgroundFPassword>
+  );
+};
+
+export default PasswordQuestionPage;
